requirements.txt
# torch >= 0.4.1
# torchvision
argparse
scipy
opencv-contrib-python
matplotlib
imageio
tqdm
tensorflow
tensorboardX
tqdm
pyyaml
imageio
imgaug
jupyter
scikit-learn

torchgeometry
torchsummary
coloredlogs
